feat(graphics3d): checkpoint in-progress 3dnextlevel2 scale-tier work
Snapshot before the navmesh voxel-baker refactor. Accumulated open-world implementation across phases 0-12: async glTF/GLB streaming with a main-thread commit queue, floating-origin camera-relative upload, Scene3D spatial index, world/terrain streaming, physics joints plus a warm-started sequential-impulse solver, navmesh/agent baselines, animation IK/additive/blend-trees/retarget, software clustered-light and occlusion baselines, KTX2/texture-asset bridge, and the openworld_slice sample with perf/hitch/GPU-smoke probes.

Plan trackers under misc/plans/game/3dnextlevel2 updated to match.

+++ src/runtime/graphics/3d/scene/rt_scene3d_internal.h
@@ -72,8 +72,8 @@ struct rt_scene3d;
 
 typedef struct {
     struct rt_scene_node3d *node;
-    float world_min[3];
-    float world_max[3];
+    double world_min[3];
+    double world_max[3];
     int32_t traversal_order;
     int8_t cullable;
 } rt_scene3d_spatial_entry;
